Moved the Edit Profile Modal into its own file and imported in profile to follow the code style of the other two modals.
